+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to the Arduino Malker Workshop extension will be documented 
 
 **Version 0.7.3**
 - Fix a problem when creating sketch on an alternate drive ([issue #30](https://github.com/thelastoutpostworkshop/arduino-maker-workshop/issues/30))
+- Add user setting to enable verbose compilation or not (default is verbose)
 
 **Version 0.7.2**
 - Arduino CLI v1.2.2